Move makedirs_if_needed from everest to ert
This function is not used exclusively by Everest, so it should be moved to a location where
it is natural to be shared between Ert and Everest.

Its logging was everest specific - which was a bug when called from the Ert side, now we do
not try to state in the logging from which side we were called from
