+++ src/ert/utils/__init__.py
@@ -1,8 +1,12 @@
 import logging
+import os
 import time
 from collections.abc import Callable
+from datetime import UTC, datetime
 from functools import wraps
 from typing import ParamSpec, TypeVar
+
+from _ert.utils import file_safe_timestamp
 
 P = ParamSpec("P")
 R = TypeVar("R")
@@ -30,3 +34,19 @@ def log_duration(
         return wrapper
 
     return decorator
+
+
+def makedirs_if_needed(path: str, roll_if_exists: bool = False) -> None:
+    if os.path.isdir(path):
+        if not roll_if_exists:
+            return
+        _roll_dir(path)  # exists and should be rolled
+    os.makedirs(path)
+
+
+def _roll_dir(old_name: str) -> None:
+    old_name = os.path.realpath(old_name)
+    timestamp = file_safe_timestamp(datetime.now(UTC).isoformat())
+    new_name = f"{old_name}__{timestamp}"
+    os.rename(old_name, new_name)
+    logging.getLogger().info(f"renamed {old_name} to {new_name}")
