+++ tests/everest/test_util.py
@@ -6,7 +6,6 @@ import pytest
 
 from ert.storage import open_storage
 from ert.storage.local_experiment import ExperimentState, ExperimentStatus
-from everest import util
 from everest.bin.utils import get_experiment_status, show_scaled_controls_warning
 from everest.strings import EVEREST
 from tests.everest.utils import (
@@ -37,57 +36,6 @@ def test_get_values(change_to_tmpdir):
     )
 
     config.environment.output_folder = rel_out_dir
-
-
-def test_makedirs(change_to_tmpdir):
-    output_dir = os.path.join("unittest_everest_output")
-    cwd = os.getcwd()
-
-    # assert output dir (/tmp/tmpXXXX) is empty
-    assert not os.path.isdir(output_dir)
-    assert len(os.listdir(cwd)) == 0
-
-    # create output folder
-    util.makedirs_if_needed(output_dir)
-
-    # assert output folder created
-    assert os.path.isdir(output_dir)
-    assert len(os.listdir(cwd)) == 1
-
-
-def test_makedirs_already_exists(change_to_tmpdir):
-    output_dir = os.path.join("unittest_everest_output")
-    cwd = os.getcwd()
-
-    # create outputfolder and verify it's existing
-    util.makedirs_if_needed(output_dir)
-    assert os.path.isdir(output_dir)
-    assert len(os.listdir(cwd)) == 1
-
-    # run makedirs_if_needed again, verify nothing happened
-    util.makedirs_if_needed(output_dir)
-    assert os.path.isdir(output_dir)
-    assert len(os.listdir(cwd)) == 1
-
-
-def test_makedirs_roll_existing(change_to_tmpdir):
-    output_dir = os.path.join("unittest_everest_output")
-    cwd = os.getcwd()
-
-    # create outputfolder and verify it's existing
-    util.makedirs_if_needed(output_dir)
-    assert os.path.isdir(output_dir)
-    assert len(os.listdir(cwd)) == 1
-
-    # run makedirs_if_needed again, verify old dir rolled
-    util.makedirs_if_needed(output_dir, True)
-    assert os.path.isdir(output_dir)
-    assert len(os.listdir(cwd)) == 2
-
-    # run makedirs_if_needed again, verify old dir rolled
-    util.makedirs_if_needed(output_dir, True)
-    assert os.path.isdir(output_dir)
-    assert len(os.listdir(cwd)) == 3
 
 
 @pytest.mark.parametrize(
